flask02/app/requirements.txt
mysql-connector-python
flasgger
gunicorn
Flask==2.3.2
gunicorn==20.1.0
